requirements.txt
biopython>=1.85
coverage>=7.5
Cython>=3.0.10
exceptiongroup>=1.2.0
iniconfig>=2.0.0
matplotlib>=3.8
mock>=5.1.0
numpy>=1.24,<2.1; python_version < "3.10"
numpy>=2.1; python_version >= "3.10"
packaging>=23.2
pluggy>=1.5.0
pytest>=8.2
pytest-cov>=5.0
pytest-mock>=3.12
tomli>=2.0.1; python_version < "3.11"
tqdm>=4.66.1
setuptools>=68
